[*] Small adjustments just to fix some weird behaviors (ball stopping on walls for example) and to make replication across network a little bit easier

+++ HaxBall/Assets/Scripts/Goal.cs
@@ -9,6 +9,6 @@ public class Goal : MonoBehaviour
 
     private void OnTriggerEnter(Collider other)
     {
-        GameController.Me.IncreaseScore((int)_playerIndex);
+        GameController.Me.IncreaseScore(_playerIndex);
     }
 }
